+++ app/server/appsmith-server/src/main/java/com/appsmith/server/controllers/ActionController.java
@@ -2,9 +2,9 @@ package com.appsmith.server.controllers;
 
 import com.appsmith.server.constants.Url;
 import com.appsmith.server.controllers.ce.ActionControllerCE;
-import com.appsmith.server.services.ActionCollectionService;
 import com.appsmith.server.services.LayoutActionService;
 import com.appsmith.server.services.NewActionService;
+import com.appsmith.server.solutions.RefactoringSolution;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -14,11 +14,11 @@ import org.springframework.web.bind.annotation.RestController;
 @Slf4j
 public class ActionController extends ActionControllerCE {
 
-    public ActionController(ActionCollectionService actionCollectionService,
-                            LayoutActionService layoutActionService,
-                            NewActionService newActionService) {
+    public ActionController(LayoutActionService layoutActionService,
+                            NewActionService newActionService,
+                            RefactoringSolution refactoringSolution) {
 
-        super(layoutActionService, newActionService);
+        super(layoutActionService, newActionService, refactoringSolution);
 
     }
