+++ app/server/appsmith-server/src/main/java/com/appsmith/server/controllers/LayoutController.java
@@ -4,6 +4,7 @@ import com.appsmith.server.constants.Url;
 import com.appsmith.server.controllers.ce.LayoutControllerCE;
 import com.appsmith.server.services.LayoutActionService;
 import com.appsmith.server.services.LayoutService;
+import com.appsmith.server.solutions.RefactoringSolution;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -14,9 +15,10 @@ import org.springframework.web.bind.annotation.RestController;
 public class LayoutController extends LayoutControllerCE {
 
     public LayoutController(LayoutService layoutService,
-                            LayoutActionService layoutActionService) {
+                            LayoutActionService layoutActionService,
+                            RefactoringSolution refactoringSolution) {
 
-        super(layoutService, layoutActionService);
+        super(layoutService, layoutActionService, refactoringSolution);
     }
 
 }
