+++ app/server/appsmith-server/src/main/java/com/appsmith/server/controllers/ce/ActionControllerCE.java
@@ -11,6 +11,7 @@ import com.appsmith.server.dtos.RefactorActionNameDTO;
 import com.appsmith.server.dtos.ResponseDTO;
 import com.appsmith.server.services.LayoutActionService;
 import com.appsmith.server.services.NewActionService;
+import com.appsmith.server.solutions.RefactoringSolution;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -40,12 +41,15 @@ public class ActionControllerCE {
 
     private final LayoutActionService layoutActionService;
     private final NewActionService newActionService;
+    private final RefactoringSolution refactoringSolution;
 
     @Autowired
     public ActionControllerCE(LayoutActionService layoutActionService,
-                              NewActionService newActionService) {
+                              NewActionService newActionService,
+                              RefactoringSolution refactoringSolution) {
         this.layoutActionService = layoutActionService;
         this.newActionService = newActionService;
+        this.refactoringSolution = refactoringSolution;
     }
 
     @PostMapping
@@ -86,7 +90,7 @@ public class ActionControllerCE {
     @PutMapping("/refactor")
     public Mono<ResponseDTO<LayoutDTO>> refactorActionName(@RequestBody RefactorActionNameDTO refactorActionNameDTO,
                                                            @RequestHeader(name = FieldName.BRANCH_NAME, required = false) String branchName) {
-        return layoutActionService.refactorActionName(refactorActionNameDTO, branchName)
+        return refactoringSolution.refactorActionName(refactorActionNameDTO, branchName)
                 .map(created -> new ResponseDTO<>(HttpStatus.OK.value(), created, null));
     }
