+++ app/server/appsmith-server/src/main/java/com/appsmith/server/controllers/ce/LayoutControllerCE.java
@@ -8,6 +8,7 @@ import com.appsmith.server.dtos.RefactorNameDTO;
 import com.appsmith.server.dtos.ResponseDTO;
 import com.appsmith.server.services.LayoutActionService;
 import com.appsmith.server.services.LayoutService;
+import com.appsmith.server.solutions.RefactoringSolution;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -30,11 +31,15 @@ public class LayoutControllerCE {
     private final LayoutService service;
     private final LayoutActionService layoutActionService;
 
+    private final RefactoringSolution refactoringSolution;
+
     @Autowired
     public LayoutControllerCE(LayoutService layoutService,
-                            LayoutActionService layoutActionService) {
+                              LayoutActionService layoutActionService,
+                              RefactoringSolution refactoringSolution) {
         this.service = layoutService;
         this.layoutActionService = layoutActionService;
+        this.refactoringSolution = refactoringSolution;
     }
 
     @PostMapping("/pages/{defaultPageId}")
@@ -74,7 +79,7 @@ public class LayoutControllerCE {
     @PutMapping("/refactor")
     public Mono<ResponseDTO<LayoutDTO>> refactorWidgetName(@RequestBody RefactorNameDTO refactorNameDTO,
                                                            @RequestHeader(name = FieldName.BRANCH_NAME, required = false) String branchName) {
-        return layoutActionService.refactorWidgetName(refactorNameDTO, branchName)
+        return refactoringSolution.refactorWidgetName(refactorNameDTO, branchName)
                 .map(created -> new ResponseDTO<>(HttpStatus.OK.value(), created, null));
     }
