+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ce/LayoutCollectionServiceCEImpl.java
@@ -24,6 +24,7 @@ import com.appsmith.server.services.AnalyticsService;
 import com.appsmith.server.services.LayoutActionService;
 import com.appsmith.server.services.NewActionService;
 import com.appsmith.server.services.NewPageService;
+import com.appsmith.server.solutions.RefactoringSolution;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
@@ -53,6 +54,7 @@ public class LayoutCollectionServiceCEImpl implements LayoutCollectionServiceCE 
 
     private final NewPageService newPageService;
     private final LayoutActionService layoutActionService;
+    private final RefactoringSolution refactoringSolution;
     private final ActionCollectionService actionCollectionService;
     private final NewActionService newActionService;
     private final AnalyticsService analyticsService;
@@ -150,13 +152,13 @@ public class LayoutCollectionServiceCEImpl implements LayoutCollectionServiceCE 
 
                     // Store the default resource ids
                     // Only store defaultPageId for collectionDTO level resource
-                    DefaultResources defaultDTOResource =  new DefaultResources();
+                    DefaultResources defaultDTOResource = new DefaultResources();
                     AppsmithBeanUtils.copyNewFieldValuesIntoOldObject(collection.getDefaultResources(), defaultDTOResource);
 
                     defaultDTOResource.setApplicationId(null);
                     defaultDTOResource.setCollectionId(null);
                     defaultDTOResource.setBranchName(null);
-                    if(StringUtils.isEmpty(defaultDTOResource.getPageId())) {
+                    if (StringUtils.isEmpty(defaultDTOResource.getPageId())) {
                         defaultDTOResource.setPageId(collection.getPageId());
                     }
                     collection.setDefaultResources(defaultDTOResource);
@@ -165,7 +167,7 @@ public class LayoutCollectionServiceCEImpl implements LayoutCollectionServiceCE 
                     DefaultResources defaults = new DefaultResources();
                     AppsmithBeanUtils.copyNewFieldValuesIntoOldObject(actionCollection.getDefaultResources(), defaults);
                     defaults.setPageId(null);
-                    if(StringUtils.isEmpty(defaults.getApplicationId())) {
+                    if (StringUtils.isEmpty(defaults.getApplicationId())) {
                         defaults.setApplicationId(actionCollection.getApplicationId());
                     }
                     actionCollection.setDefaultResources(defaults);
@@ -309,7 +311,7 @@ public class LayoutCollectionServiceCEImpl implements LayoutCollectionServiceCE 
                     return actionUpdatesFlux
                             .then(actionCollectionService.update(branchedActionCollection.getId(), branchedActionCollection))
                             .then(branchedPageIdMono)
-                            .flatMap(branchedPageId -> layoutActionService.refactorName(branchedPageId, layoutId, oldName, newName));
+                            .flatMap(branchedPageId -> refactoringSolution.refactorName(branchedPageId, layoutId, oldName, newName));
                 })
                 .map(responseUtils::updateLayoutDTOWithDefaultResources);
     }
@@ -590,20 +592,20 @@ public class LayoutCollectionServiceCEImpl implements LayoutCollectionServiceCE 
                 .map(responseUtils::updateCollectionDTOWithDefaultResources)
                 .zipWith(newPageService.findById(pageId, MANAGE_PAGES),
                         (branchedActionCollection, newPage) -> {
-                        //redundant check
-                    if (newPage.getUnpublishedPage().getLayouts().size() > 0 ) {
-                        // redundant check as the collection lies inside a layout. Maybe required for testcases
-                        branchedActionCollection.setErrorReports(newPage.getUnpublishedPage().getLayouts().get(0).getLayoutOnLoadActionErrors());
-                    }
+                            //redundant check
+                            if (newPage.getUnpublishedPage().getLayouts().size() > 0) {
+                                // redundant check as the collection lies inside a layout. Maybe required for testcases
+                                branchedActionCollection.setErrorReports(newPage.getUnpublishedPage().getLayouts().get(0).getLayoutOnLoadActionErrors());
+                            }
 
-                    return branchedActionCollection;
-                });
+                            return branchedActionCollection;
+                        });
     }
 
     @Override
     public Mono<LayoutDTO> refactorAction(RefactorActionNameInCollectionDTO refactorActionNameInCollectionDTO) {
         // First perform refactor of the action itself
-        final Mono<LayoutDTO> layoutDTOMono = layoutActionService
+        final Mono<LayoutDTO> layoutDTOMono = refactoringSolution
                 .refactorActionName(refactorActionNameInCollectionDTO.getRefactorAction())
                 .cache();
 
@@ -642,7 +644,7 @@ public class LayoutCollectionServiceCEImpl implements LayoutCollectionServiceCE 
                 });
 
         Mono<String> branchedCollectionIdMono = StringUtils.isEmpty(branchName)
-                ?  Mono.just(defaultActionCollection.getId())
+                ? Mono.just(defaultActionCollection.getId())
                 : actionCollectionService
                 .findByBranchNameAndDefaultCollectionId(branchName, defaultActionCollection.getId(), MANAGE_ACTIONS)
                 .map(actionCollection -> {
