+++ app/server/appsmith-server/src/main/java/com/appsmith/server/solutions/RefactoringSolution.java
@@ -1,0 +1,6 @@
+package com.appsmith.server.solutions;
+
+import com.appsmith.server.solutions.ce.RefactoringSolutionCE;
+
+public interface RefactoringSolution extends RefactoringSolutionCE {
+}
